FRONT-1011: New toaster, new toasts

+++ src/features/alias/sagas/set.saga.ts
@@ -1,72 +1,77 @@
+import { ToastType } from '$/components/Toast'
 import { AliasAction } from '$/features/alias'
+import toast from '$/features/toaster/helpers/toast'
 import { Address } from '$/types'
 import db from '$/utils/db'
 import handleError from '$/utils/handleError'
 import isBlank from '$/utils/isBlank'
-import { error } from '$/utils/toaster'
 import { takeEvery } from 'redux-saga/effects'
 
 async function create(owner: Address, address: Address, value: string) {
-    try {
-        const now = Date.now()
+    const now = Date.now()
 
-        await db.aliases.add({
-            createdAt: now,
-            updatedAt: now,
-            owner: owner.toLowerCase(),
-            address: address.toLowerCase(),
-            alias: value,
-        })
-    } catch (e) {
-        error(`Failed to create a nickname.`)
-
-        throw e
-    }
+    await db.aliases.add({
+        createdAt: now,
+        updatedAt: now,
+        owner: owner.toLowerCase(),
+        address: address.toLowerCase(),
+        alias: value,
+    })
 }
 
 async function destroy(owner: Address, address: Address) {
-    try {
-        await db.aliases
-            .where({ owner: owner.toLowerCase(), address: address.toLowerCase() })
-            .delete()
-    } catch (e) {
-        error(`Failed to remove a nickname.`)
-
-        throw e
-    }
+    await db.aliases.where({ owner: owner.toLowerCase(), address: address.toLowerCase() }).delete()
 }
 
 async function update(owner: Address, address: Address, value: string) {
-    let numModded = 0
-
-    try {
-        numModded = await db.aliases
-            .where({ owner: owner.toLowerCase(), address: address.toLowerCase() })
-            .modify({ alias: value, updatedAt: Date.now() })
-    } catch (e) {
-        error(`Failed to update a nickname.`)
-
-        throw e
-    }
-
-    return numModded
+    return db.aliases
+        .where({ owner: owner.toLowerCase(), address: address.toLowerCase() })
+        .modify({ alias: value, updatedAt: Date.now() })
 }
 
 export default function* set() {
     yield takeEvery(AliasAction.set, function* ({ payload: { owner, address, value } }) {
         try {
             if (isBlank(value)) {
-                yield destroy(owner, address)
+                try {
+                    yield destroy(owner, address)
+                } catch (e) {
+                    yield toast({
+                        title: 'Failed to remove a nickname',
+                        type: ToastType.Error,
+                    })
+
+                    throw e
+                }
+
                 return
             }
 
-            const numModded: number = yield update(owner, address, value)
+            try {
+                const numModded: number = yield update(owner, address, value)
 
-            if (numModded > 0) {
-                return
+                if (numModded > 0) {
+                    return
+                }
+            } catch (e) {
+                yield toast({
+                    title: 'Failed to update a nickname',
+                    type: ToastType.Error,
+                })
+
+                throw e
             }
 
-            yield create(owner, address, value)
+            try {
+                yield create(owner, address, value)
+            } catch (e) {
+                yield toast({
+                    title: 'Failed to create a nickname',
+                    type: ToastType.Error,
+                })
+
+                throw e
+            }
         } catch (e) {
             handleError(e)
         }
